backend/requirements.txt
fastapi
uvicorn[standard]
langchain-cohere
python-dotenv
pydantic